requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
argcomplete==3.6.2
attrs==25.3.0
bcrypt==4.1.2
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colored==2.3.0
cryptography==45.0.4
data==0.4
decorator==5.2.1
dotenv==0.9.9
et_xmlfile==2.0.0
Flask==3.1.1
fpdf==1.7.2
frozenlist==1.7.0
funcsigs==1.0.2
google==3.0.0
greenlet==3.2.4
idna==3.10
inotify==0.2.10
itsdangerous==2.2.0
Jinja2==3.1.6
login==0.0.6
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.6.3
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
nose==1.3.7
numpy==2.3.1
openpyxl==3.1.5
paho-mqtt==2.1.0
pandas==2.3.0
pdfminer.six==20250506
pdfs==0.1.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
requests==2.32.4
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.43
twilio==9.7.0
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
Whoosh==2.7.4
wikipedia==1.4.0
yarl==1.20.1
